refactor: move some webserver-exclusive functions behind webserver gate
