+++ freqtrade/rpc/api_server/webserver.py
@@ -124,6 +124,7 @@ class ApiServer(RPCHandler):
         from freqtrade.rpc.api_server.api_pairlists import router as api_pairlists
         from freqtrade.rpc.api_server.api_v1 import router as api_v1
         from freqtrade.rpc.api_server.api_v1 import router_public as api_v1_public
+        from freqtrade.rpc.api_server.api_webserver import router as api_webserver
         from freqtrade.rpc.api_server.api_ws import router as ws_router
         from freqtrade.rpc.api_server.deps import is_webserver_mode
         from freqtrade.rpc.api_server.web_ui import router_ui
@@ -135,6 +136,11 @@ class ApiServer(RPCHandler):
             api_v1,
             prefix="/api/v1",
             dependencies=[Depends(http_basic_or_jwt_token)],
+        )
+        app.include_router(
+            api_webserver,
+            prefix="/api/v1",
+            dependencies=[Depends(http_basic_or_jwt_token), Depends(is_webserver_mode)],
         )
         app.include_router(
             api_backtest,
